[refactor] 키워드 관련 훅 keyword 디렉토리로 이동
